Add tile_matrix grid config and placeholder for WMTS sources
Currently MapProxy can be configured with a suitable grid and tile
source in order to treat a WMTS as a source. The source url is a GetTile
request with placeholders for tilecol (x) and tilerow (y) plus the
tilematrix. The tilematrix is commonly substituted with the tile zoom
level (z) which in practice works for the majority of WMTS servers but
not all: The tilematrix value should really be the <ows:Identifier>
value for the corresponding <TileMatrix> defined in the WMTS
GetCapabilities response.  The WMTS 1.0.0 spec states that a TileMatrix
Identifier is a "CodeType" and includes an example capabilities document
with TileMatrix Identifiers such as "2g", "1g" and "30m".

This commit extends grid config to include an optional tile_matrix
property which is a list of TileMatrix Identifiers in the same order as
the grid resolutions; a new "tile_matrix" substituion token is also made
available for use in a tile source url.